code quality - cut unnecessary calls to db, moved utilities into utils folder
